[COD-53] Search logic moved to the service and allowed for unauthorised users see course list
